JR: Change 64 bit ints from long long to long

+++ JR/Types.h
@@ -4,12 +4,12 @@
 using i8 = char;
 using i16 = short;
 using i32 = int;
-using i64 = long long;
+using i64 = long;
 
 using u8 = unsigned char;
 using u16 = unsigned short;
 using u32 = unsigned int;
-using u64 = unsigned long long;
+using u64 = unsigned long;
 
 using f32 = float;
 using f64 = double;
